+++ backend/app/models/__init__.py
@@ -4,5 +4,14 @@ from .library import MediaLibrary
 from .progress import PlaybackProgress
 from .scan_job import ScanJob
 from .actor_photo import ActorPhoto
+from .invite_code import InviteCode
 
-__all__ = ["User", "MediaItem", "MediaLibrary", "PlaybackProgress", "ScanJob", "ActorPhoto"]
+__all__ = [
+    "User",
+    "MediaItem",
+    "MediaLibrary",
+    "PlaybackProgress",
+    "ScanJob",
+    "ActorPhoto",
+    "InviteCode",
+]